requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
plyfile==1.1
pooch==1.8.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyvista==0.45.1
PyYAML==6.0.2
requests==2.32.3
scooby==0.10.0
six==1.16.0
trame-client==3.5.0
trame-vtk==2.8.12
typing_extensions==4.12.2
urllib3==2.2.3
vtk==9.3.1
zipp==3.20.2
